pyproject.toml
[project]
name = "Toodles"
version = "0.9a"
description = "Read weights from serial port and write to a file."
authors = [
    {name = "Francisco José Rodríguez Bogado", email = "bogado@qinn.es"},
]
dependencies = [
    "pyserial",
    "fire",
]
requires-python = ">=3.9"
license = {text = "GPLv3"}

[project.urls]
homepage = ""

[tool]
[tool.pdm]

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
